Поменял структуру проекта, немного изменил основные файлы, добавил tests.cpp
