+++ CMakeLists.txt
@@ -1,4 +1,9 @@
 cmake_minimum_required(VERSION 3.24.2)
-project(testproj CXX)
 
-add_executable(${PROJECT_NAME} main.cpp)
+project(
+    testproj
+    VERSION 1.0.0
+    DESCRIPTION "Test project")
+
+add_subdirectory(candleLib)
+add_subdirectory(candleApp)
