+++ candleLib/include/tests/tests.h
@@ -1,0 +1,4 @@
+#ifndef CANDLELIB_INCLUDE_TESTS_TESTS_H_
+#define CANDLELIB_INCLUDE_TESTS_TESTS_H_
+
+#endif
