+++ candleLib/src/tests.cpp
@@ -1,0 +1,1 @@
+#include "tests/tests.h"
